moved some classes in UtiLITI. built in a Focus in the MapComponent (the scrolling, zooming, editing etc. only works when not hovering GUIcomponents). the "indestructible" flag as well as selecting the spritesheets of props is now also possible. (The PropAnimationController however does not assign new animations when the Prop is indestructible.) DropdownListField now has a getSelectedIndex and a getSelectedObject. before, it only had a method to retrieve the selection index. rendering the collisionBoxes in the MapComponent now. ImageComponents now set their Image size each time their total size is changed.
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@2239 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/graphics/IGuiComponent.java
@@ -39,13 +39,13 @@ public interface IGuiComponent extends IRenderable {
 
   void setDimension(double width, double height);
 
-  public void setHeight(int height);
+  public void setHeight(double height);
 
   void setPosition(double x, double y);
 
   void setPosition(Point2D newPosition);
 
-  public void setWidth(int width);
+  public void setWidth(double width);
 
   /**
    * Suspend.
